requirements.txt
streamlit==0.89.0
Pillow==8.2.0
pandas==1.3.3
numpy==1.19.5
scikit-learn==1.0